requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
GeoAlchemy2==0.11.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
packaging==21.3
psycopg2-binary==2.9.3
PyMySQL==1.0.2
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
requests==2.27.1
Shapely==1.8.1.post1
SQLAlchemy==1.4.32
timeago==1.0.15
urllib3==1.26.8
Werkzeug==2.0.3
